src/main/java/cz/cvut/fel/ear/stepavi2_havriboh/main/security/AuthenticationFailure.java
package cz.cvut.fel.ear.stepavi2_havriboh.main.security;

import com.fasterxml.jackson.databind.ObjectMapper;
import cz.cvut.fel.ear.stepavi2_havriboh.main.security.model.LoginStatus;
import cz.cvut.fel.ear.stepavi2_havriboh.main.utils.Constants;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.web.authentication.AuthenticationFailureHandler;

import java.io.IOException;

/**
 * Returns info about authentication failure.
 *
 * Differs from default implementation in that it returns a custom JSON response.
 */
public class AuthenticationFailure implements AuthenticationFailureHandler {

    private static final Logger LOG = LoggerFactory.getLogger(AuthenticationFailure.class);

    private final ObjectMapper mapper;

    public AuthenticationFailure(ObjectMapper mapper) {
        this.mapper = mapper;
    }

    @Override
    public void onAuthenticationFailure(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse,
                                        AuthenticationException e) throws IOException {
        LOG.debug("Login failed for user {}.", httpServletRequest.getParameter(Constants.USERNAME_PARAM));
        final LoginStatus status = new LoginStatus(false, false, null, e.getMessage());
        mapper.writeValue(httpServletResponse.getOutputStream(), status);
    }
}
